Revert "Change repository structure"
This reverts commit e8dd0ba5322d536dd8d3046de955a712553d3633.
